Task 0

+++ 0x17-doubly_linked_lists/0-print_dlistint.c
@@ -9,7 +9,7 @@
 
 size_t print_dlistint(const dlistint_t *h)
 {
-	int i = 0;
+	size_t i = 0;
 	dlistint_t *head = NULL;
 
 	h = head;
@@ -20,5 +20,6 @@ size_t print_dlistint(const dlistint_t *h)
 		h = h->next;
 		i++;
 	}
+
 	return (i);
 }
